requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
azure-identity==1.15.0
msgraph-sdk==1.0.0
neo4j==5.17.0
langchain==0.1.7
langchain-openai==0.0.6
langchain-community==0.0.20
python-dotenv==1.0.1
pydantic==2.6.1